tests/wpt/tests/html/canvas/offscreen/shadows/2d.shadow.image.transparent.2.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.shadow.image.transparent.2</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.shadow.image.transparent.2</h1>
<p class="desc">Shadows are not drawn for transparent parts of images</p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 50, 50);
  ctx.fillStyle = '#f00';
  ctx.fillRect(50, 0, 50, 50);
  ctx.shadowOffsetY = 50;
  ctx.shadowColor = '#0f0';
  var response = await fetch('/images/redtransparent.png')
  var blob = await response.blob();
  var img = await createImageBitmap(blob);
  ctx.drawImage(img, 50, -50);
  ctx.shadowColor = '#f00';
  ctx.drawImage(img, -50, -50);

  _assertPixel(canvas, 25,25, 0,255,0,255);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 75,25, 0,255,0,255);

}, "Shadows are not drawn for transparent parts of images");
</script>
